Extract dashboard ROS payload types
